pymi/requirements-test.txt
flake8
mypy==0.560
pylint
bandit==1.4.0
# For ex8
PyYAML>=3.11
